refactor: simplify Layout component by removing unused hooks and add useCaptureInput hook for input handling
